pyproject.toml
[tool.poetry]
name = "ha_lantuya"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
homeassistant = "^2021.6.6"
pytest-asyncio = "^0.15.1"
autoflake = "^1.4"
pycryptodome = "^3.10.1"
autopep8 = "^1.5.7"
black = {version = "^21.7b0", allow-prereleases = true}
tzdata = "^2021.1"
pylama = "^7.7.1"
pylint = "^2.9.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "debug"

